Added initial go_db project
The go db is a DBMS implementation written entirely in Go.
Currently written initial draft of the Database structures in storage
and started implementing basic IO operations such as reading and writing
from "data blocks" which are unit of storage from which all DB
structures will be composed of.

Currently the IO logic is written so that every operation done will open
the DB file for exclusive usage, which obviously isn't very efficient.
We can optimize this behaviour later by creating locks protecting
specific parts of the DB file (for example, one can read the content of
table A while another thread concurrently changes the contents of table
B, as this shouldn't cause a problem).