Implemented deleteRecord
Currently only deletes based in the record index (which will be used as
the low level deletion implementation in the future).
Deletion only overrides the corresponding bit in the table bitmap.
In order to decrease the size of a table after many deletions we would
have to perform a DB check and move many structures around...