condition implementation
Added go_db/condition.go, which implements logic required to run a set
of conditions on a single record.
This set of condition is represented the conditionNode struct, which
includes an operator and two optional operands.
When "checkAllConditions" is executed against a Record we run through
the conditionNode graph, recursively calling checkAllConditions on all
connected nodes, and applying the appropriate operators on their
results.