go_db: Basic Connection and Cursor implementation
Added the Cursor and Connection structs which are public.
Users can use the Connection struct to establish a connection to the DB,
and from it to obtain a Cursor struct with which they can perform
queries on the DB.
The added e2e test details how to use the API.
Currently only select queries are exposed.