go_db: Added initial version of DB server
The server will communicate with a UI component that is yet to be
written to allow comfortable query of the DB's data.
Currently there is a bit of a mess in the packages due to my lack of
knowledge, but we can fix that in the future.
I've also added a sample client which exposes the bare bones
functionality to allow us to query the DB.

The only crucial logic that is missing at the moment is a response
containing the records queried.

We should probably implement a serialization method for that and expose
that in go_db itself instead of the server (which should be as simple as
possible).