go_db: Implementation of record iteration using goroutines.
The new implementation allows the framework to perform some operation on
every record inside a table and return the result of this operation.

For example - the select query requires the framework to perform the
following for every record in the table:
1. Check if the "where" condition is true for this record.
2. Create a new record according to the columns (and their order) in the
   select clause.
3. Return this new record.

Now all this operation can be performed concurrently in different
goroutines.