go_db: expansion of free block bitmap beyond a single data block
Until now the size of the free blocks bitmap was capped by a single data
block, since we didn't have proper implementation of its expansion
(since that will require us to allocate a new data block, which in turn
requires the bitmap we're currently modifying).

The way this is done is simply by detecting we have reached passed the
current allocated size for the bitmap (the new requested is both bigger
than the current size of the bitmap as well as aligned with the size of
a single data block - meaning the corresponding bit in the bitmap will
reside in a new data block).