go_db: Fixed deletion-then-insertion bug
Add a bug when inserting a record in an index in which a previous record
used to reside (this can happen since when we delete a record we only
mark it as deleted in the bitmap and not actually delete it).

The issue was we miscalculated the size of a record as a single db
pointer, although each record is the size of <DB_POINTER>*<amount of
columns>.

Added a test for this case as well, which adds 10K records after
deleting previous records.
This also test our bitmap expension feature, and the general robustness
of the code.