b_tree: Implement remove
Implemented BTree's remove method, which removes an item from the tree.
In order to keep the tree balaned, we do not keep small nodes (so that
when the tree's item count shrinks so will the tree's height).
In order to enforce that, in the removal process we force every node in
the path to the leaf node containing the item to remove to contain at
least maximumSize/2 + 1 items.

If a node does not contain the minimum amount of items, we enlarge it by
the following logic:
1. If one of his neighbors is big enough (more than maximumSize/2
   items), steal an item from it.
2. If step #1 failed, merge the node with one of its brothers.
3. If the new merged node has no brothers, swap it with its father and
   delete its father.
   Step #3 can only occur at the root, due to the enforcement that every
   node will have at least maximumSize/2+1 items (except for root).