b_tree: Added persistency api layer and serialization interface
Added a layer of abstraction in the process of persisting BTree nodes.
The persistency is composed of two layers:
The first is a concrete API implemented by the b_tree package which
serializes the objects into array bytes.
The second is an abstract interface supplied by the caller which
actually persists the data into some persistent memory.

An in-memory implementation of a persistency store is implemented in
in_memory_persistency.go.