b_tree: Improve persistencyInterface to allow integration with callers
Added the required prerequisites so that other packages could interact
properly with the b_tree module.
Specifically, exposed BTreePointer (as the caller must be aware of this
type in order to supply it back from its PersistencyInterface objet),
and added the required method "RootPointer" to the interface.

The RootPointer method is necessary for the b_tree module to work since
we need a way to grab the data of our root pointer when we initialize
the module.
Since we have no way to persist data on our own, we are entirely reliant
on the caller to supply some interface by which we can load our root
data structure into memory.
That API is through the RootPointer method, which simply returns a
pointer we can use to Persist/Load our root data structure.