b_tree: Enforce no key duplicates are seen in BTree
Duplication of key is not supported in the B-Tree as this could leave to
unrecoverable edge-cases.
For example, if we have N+1 items with key=X, and NodeSize=N, then we
are bound to have the same key in multiple leaves.
That would mean we will have no way of guarantee in which leaf the item
actually resides.

If we go one step further, if we allow key duplication, the entire tree
can contain only a single key (with lots of different pointers), then
removal, insertion and search will all take O(n).

Therefore, it is crucial we don't support this feature.