go_db: support JOINs in record iteration
Added support for JOINs (which are only supported in SELECT queries) in
the DB's record iteration module.
JOINs are done by performing a cartesian product on all mentioned
tables.
Currently we retrieve all the data of the record in each table and merge
them together into one big record.
Only in the final stage we perform the "select" operation and remove all
unnecessary fields.