go_db: changed "heavy" tests to use in-memory buffer instead of file
Some tests take a long time to run on a file, and since we don't handle
concurrency properly running all the tests at once causes them to fail
opening the file, and therefore always fail.
By moving to in-memory buffer we guarantee all of those tests run
extremely fast and finish in less than a second.