go_db: add provenance to DB
Added the concept of provenance - metadata about the origin of records.
The provenance's data is generated during the initialization of the
openDB object (which is created for each instance of a connected user),
which then adds its data into each record.

The provenance otherwise functions similarly to a standard column,
except it has its own member in the Record object and in the
tableScheme.

Despite being identical on all tables of the same DB, we still keep the
column headers of the provenance fields in each table.
This is requires only a small overhead of space and allows us to easily
index the provenance.