go_db: index provenance columns
Added the support to index the provenance columns.
The provenance index works a bit differently than a normal column index,
that is because the score of a given record's provenance (which is the
key of its index) is not unique among records.

For this reason we don't keep the pointer to the actual underlying
record as a value in the index's BTree structure, instead, for each
provenance score we save a pointer to another BTree, in which the keys
are the offsets of the records with this provenance score, and the
values are dummy values.

Therefore, in order to iterate over all records in order of their
provenance we need to do the following:
Iterate the items in the provenance index.
For each item:
1. Load the pointer as a BTree T.
2. Iterate T, for every item:
2.1. The key of the item is the record offset in its corresponding
  table.

Note that we'll get the records in a meaningless order in step 2.1, but
that is fine because they all have the same provenance score and are
therefore orderless in the eyes of the iterator.

Side note:
This change required us to modify record iteration, especially in the
areas surrounding modifications (deletion/update).
This commit merely introduces provenance indexing and does not supply
full support for provenance propogation in all the appropriate
operations (for example - we should expect UPDATE to override the
provenance or at the very least perform a "*" operation in the
provenance domain).