go_db: Provenance multiplication support
Added support for provenance multiplication (which occurs on record
JOIN).
The behavior of the provenance fields when records are joined is to
perform a logical "*" operation between all the provenance fields of the
same type.

For this to work we had to modify the structure of the ProvenanceField,
which now contains either a Value (indicating this is a "simple" field)
or an operator and a list of pointer to other provenances (the
operands).

This way we can chain together complex arithmetic expressions of
provenances using parentheses (represented by the provenances fields
that are the "nodes" in the tree), multiplication/addition operators
(which are represented by the "operator" field in the ProvenanceField
structure) and atomic values (which are the "simple" provenance fields).

Currently we only support multiplication, addition (which is performed
on SELECT) will be added in a later commit.