go_db: Provenance aggregation support
Added the capability to properly calculate the score of a record's
provenance, even when the provenance isn't "simple" (it is the result of
JOINs/SELECT and therefore contains multiple inner provenance fields).

We calculate the overall provenance score by defining two aggregation
functions - one for the "*" operator of the provenance field, and
another for the "+" operator.

Those functions are passed for each connection (preset in the OpenDB
structure) and are used dynamically by the ProvenanceField object to
determine its score.