go_db: expose provenance in Connect
Exposed optional provenance parameters in the "Connect" function which
is used to get a connection to the DB.

As part of this change we've modified all table APIs (readAllRecords,
deleteRecord, etc) which accepted an unopened database as an argument so
that they'll require an open database from now.

This will help us consolidate the ways users have to access the DB and
ensure we will hold the correct provenance at hand when performing
actions on the DB's datastructures.