go_db: Fix order of fields in select result (bugfix)
Fixed a bug in which the order of the retrieved fields (in SELECT query)
would always be their order in the joint table scheme, instead of the
order in which they appear in the SELECT query (which is the expected
behaviour).

Also fixed a bug in which non-existent column names could be mentioned
in the SELECT clause and no error would occur (they will simply vanish
from the result).