go_db: Fix bug in index iterator on non-prov index (bugfix)
The implementation of the index iterator only handled non-unique index
keys.
However, when the key of the index is unique, we need to treat the BTree
pair differently (instead of treating the pointer as a DB pointer to
another BTree, we can just use it as the record offset immediately).

Added a test to verify this behavior as well.