go_db: Added "best" keyword to select queries
The new keyword can be used after the "order by" keyword and its format
is:
"best <integer>".

Once applied, the db engine will use the provenance top K retrieve
function instead of iterating the tables normally.

As part of this change we've also changed the flow of adding new tables
in the database.
Now whenever a new table is added, if provenance is on then all of its
provenance columns will be automatically indexed.