go_db: Change conn+auth scores to be more reasonable
Since our score is better the smaller it is we made a change to how the
scores are calculated for authentication/connection provenances.

Now bigger passwords will award the user with better (smaller)
provenances and the only part of the connection IP that will be used
for the provenance field is the most significant byte of it.